Fix check on host side to know whether we already have the data about all the SPs

+++ src/dpu_offload_event_channels.c
@@ -2206,7 +2206,7 @@ static dpu_offload_status_t sp_data_recv_cb(struct dpu_offload_ev_sys *ev_sys, e
     assert(!ev_sys->econtext->engine->on_dpu);
 
     // If we already have the data, do nothing.
-    if (econtext->engine->host.total_num_sps != UINT64_MAX)
+    if (econtext->engine->host.total_num_sps != SIZE_MAX)
         return DO_SUCCESS;
 
     assert(econtext->engine->host_dpu_data_initialized == true);
